Unify NodePoolNodeValues type
