+++ libs/ui-lib/lib/cim/components/Hypershift/modals/NodePoolForm.tsx
@@ -15,22 +15,13 @@ import { useTranslation } from '../../../../common/hooks/use-translation-wrapper
 import { AgentK8sResource } from '../../../types';
 import { getAgentsForSelection, getVersionFromReleaseImage } from '../../helpers';
 import NodePoolAgentsForm from '../forms/NodePoolAgentsForm';
-import { AgentMachineK8sResource, HostedClusterK8sResource, NodePoolK8sResource } from '../types';
+import {
+  AgentMachineK8sResource,
+  HostedClusterK8sResource,
+  NodePoolFormValues,
+  NodePoolK8sResource,
+} from '../types';
 import { getNodepoolAgents } from '../utils';
-
-export type NodePoolFormValues = {
-  nodePoolName: string;
-  agentLabels: {
-    key: string;
-    value: string;
-  }[];
-  count: number;
-  autoscaling: {
-    minReplicas: number;
-    maxReplicas: number;
-  };
-  useAutoscaling: boolean;
-};
 
 type NodePoolFormProps = {
   agents: AgentK8sResource[];
